Display current weather information for last searched for location on loading
